requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
flask-restx==1.3.0
python-dotenv==1.0.0
Flask-restful
Flask-marshmallow
marshmallow-SQLAlchemy